refactor(runtime): share queue execution across web and tauri
