+++ crates/gpt-image-2-runtime/src/lib.rs
@@ -1,15 +1,27 @@
 #![allow(unused_imports)]
 
 mod batch_payloads;
+mod command;
+mod edit_runner;
+mod generate_runner;
 mod job_paths;
 mod job_records;
 mod provider_capabilities;
 mod queue_events;
+mod queue_executor;
+mod queue_jobs;
 mod queue_types;
+mod streaming;
 
 pub use batch_payloads::*;
+pub use command::*;
+pub use edit_runner::*;
+pub use generate_runner::*;
 pub use job_paths::*;
 pub use job_records::*;
 pub use provider_capabilities::*;
 pub use queue_events::*;
+pub use queue_executor::*;
+pub use queue_jobs::*;
 pub use queue_types::*;
+pub use streaming::*;
